payment

+++ database/migrations/2018_06_11_194201_create_monetary_deatil.php
@@ -24,11 +24,11 @@ class CreateMonetaryDeatil extends Migration
             $table->integer('deposit')->nullable();
             $table->integer('second_payment')->nullable();
             $table->integer('final_payment')->nullable();
-            $table->integer('half_title')->nullable();
-            $table->integer('half_land_agreement')->nullable();
-            $table->integer('half_build_agreement')->nullable();
-            $table->integer('half_stamp_duty')->nullable();
-            $table->integer('half_reg_fee')->nullable();
+            $table->float('half_title')->nullable();
+            $table->float('half_land_agreement')->nullable();
+            $table->float('half_build_agreement')->nullable();
+            $table->float('half_stamp_duty')->nullable();
+            $table->float('half_reg_fee')->nullable();
             $table->integer('inc_cost')->nullable();
             $table->integer('maintenance_expense')->nullable();
             $table->integer('identification_fee')->nullable();
